Refactor XML serialization into CycloneDX.Xml library
